# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: XAI Blockchain API
  version: 1.1.0
  description: 'Production REST API for the XAI blockchain node.


    All endpoints require authenticated transport in production. Use the summaries
    and schemas in this document to integrate wallets, explorers, trading engines,
    and operational tooling.'
  contact:
    name: XAI Support
    url: https://xai-blockchain.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:5000
  description: Local development
- url: https://testnet-api.xai-blockchain.io
  description: Testnet
- url: https://api.xai-blockchain.io
  description: Production
...
paths:
...
  /blocks:
    get:
      tags:
      - Blockchain
      summary: List blocks
      description: Returns the latest blocks with pagination.
      parameters:
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  blocks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Block'
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /blocks/{index}:
    get:
      tags:
      - Blockchain
      summary: Get block by index
      description: Fetch a block using its height/index with cache-friendly headers.
      parameters:
      - $ref: '#/components/parameters/BlockIndexParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /block/{block_hash}:
    get:
      tags:
      - Blockchain
      summary: Get block by hash
      description: Fetch a block using its canonical hash.
      parameters:
      - $ref: '#/components/parameters/BlockHashParam'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
  /block/receive:
    post:
      tags:
      - Blockchain
      summary: Accept peer block
      description: Receives a block envelope signed by a peer and attempts to add
        it to the chain.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeerBlockInput'
      responses:
        '200':
          description: Request succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  height:
                    type: integer
                required:
                - success
        '400':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Maximum number of records to return.
    OffsetParam:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Pagination offset.
...
    BlockIndexParam:
      name: index
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
    BlockHashParam:
      name: block_hash
      in: path
      required: true
      schema:
        type: string
        pattern: ^[0-9a-fA-Fx]+$
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
    Transaction:
      type: object
      properties:
        txid:
          type: string
        sender:
          type: string
        recipient:
          type: string
        amount:
          type: number
        fee:
          type: number
        nonce:
          type: integer
        timestamp:
          type: number
    Block:
      type: object
      properties:
        index:
          type: integer
        timestamp:
          type: number
        previous_hash:
          type: string
        hash:
          type: string
        merkle_root:
          type: string
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
...
    PeerBlockInput:
      type: object
      required:
      - header
      - transactions
      properties:
        header:
          type: object
        transactions:
          type: array
          items:
            type: object
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- api_key: []
- bearer_auth: []